fix(gateway): don't set x-forwarded-for for loopback client IPs
getClientIp() returns "127.0.0.1" for loopback peers (never undefined),
so the else branch that deletes x-forwarded-for was dead code. The
runtime's stricter check (reject any x-forwarded-for in bare-metal mode,
added in #22591) then 403'd legitimate local bootstrap requests because
the gateway was forwarding 127.0.0.1 as x-forwarded-for.

Fix: check isLoopbackAddress(clientIp) and skip setting the header for
loopback peers in both proxy handlers. Extract isLoopbackAddress into a
shared gateway utility.
