+++ gateway/src/http/routes/browser-relay-websocket.ts
@@ -4,6 +4,7 @@ import {
 } from "../../auth/token-exchange.js";
 import type { GatewayConfig } from "../../config.js";
 import { getLogger } from "../../logger.js";
+import { isLoopbackAddress } from "../../util/is-loopback-address.js";
 
 const log = getLogger("browser-relay-ws");
 
@@ -46,28 +47,6 @@ function isPrivateNetworkPeer(
   const ip = server.requestIP(req);
   if (!ip) return false;
   return isPrivateAddress(ip.address);
-}
-
-/**
- * Stricter loopback-only check: accepts only 127.0.0.0/8 and ::1.
- * Use this instead of isPrivateNetworkPeer for endpoints that must be
- * restricted to the local machine (e.g. token minting).
- */
-function isLoopbackAddress(addr: string): boolean {
-  const v4Mapped = addr.match(/^::ffff:(\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})$/i);
-  const normalized = v4Mapped ? v4Mapped[1] : addr;
-
-  if (normalized.includes(".")) {
-    const parts = normalized.split(".").map(Number);
-    if (
-      parts.length !== 4 ||
-      parts.some((p) => Number.isNaN(p) || p < 0 || p > 255)
-    )
-      return false;
-    return parts[0] === 127;
-  }
-
-  return normalized.toLowerCase() === "::1";
 }
 
 export function isLoopbackPeer(
