+++ gateway/src/http/routes/runtime-proxy.ts
@@ -6,6 +6,7 @@ import {
 import type { GatewayConfig } from "../../config.js";
 import { fetchImpl } from "../../fetch.js";
 import { getLogger } from "../../logger.js";
+import { isLoopbackAddress } from "../../util/is-loopback-address.js";
 import { stripHopByHop } from "../../util/strip-hop-by-hop.js";
 
 const log = getLogger("runtime-proxy");
@@ -87,9 +88,10 @@ export function createRuntimeProxyHandler(config: GatewayConfig) {
 
     // Inject the real client IP so the runtime can rate-limit per-user,
     // overwriting any client-supplied value to prevent spoofing.
-    // When clientIp is undefined (loopback peer), strip any client-supplied
-    // header to prevent forged forwarded-for values from reaching the runtime.
-    if (clientIp) {
+    // Strip the header for loopback peers: the runtime rejects any request
+    // with x-forwarded-for in bare-metal mode, and forwarding 127.0.0.1
+    // conveys no useful information.
+    if (clientIp && !isLoopbackAddress(clientIp)) {
       reqHeaders.set("x-forwarded-for", clientIp);
     } else {
       reqHeaders.delete("x-forwarded-for");
